@@ -2,10 +2,10 @@
   "name": "google-fonts-plugin",
   "version": "4.1.0",
   "description": "Webpack plugin that downloads fonts from Google Fonts and encodes them to base64.",
-  "main": "dist/index.js",
+  "main": "dist/Plugin.js",
   "scripts": {
     "prepublishOnly": "npm test && npm run build",
-    "pretest": "standard --fix",
+    "pretest": "standard *.ts",
     "test": "jest --coverage --env=node",
     "posttest": "codecov",
     "test-watch": "jest --watch --env=node",
@@ -28,29 +28,31 @@
     "url": "https://github.com/SirPole/google-fonts-plugin/issues"
   },
   "homepage": "https://github.com/SirPole/google-fonts-plugin#readme",
-  "babel": {
-    "presets": [
-      [
-        "@babel/preset-env",
-        {
-          "targets": {
-            "node": "10"
-          }
-        }
-      ]
-    ],
-    "plugins": [
-      "@babel/plugin-proposal-class-properties"
-    ]
-  },
   "standard": {
-    "parser": "babel-eslint",
+    "parser": "@typescript-eslint/parser",
+    "plugins": [
+      "@typescript-eslint/eslint-plugin"
+    ],
     "env": {
       "node": true,
       "commonjs": true,
       "es6": true,
       "jest": true
     }
+  },
+  "jest": {
+    "roots": [
+      "<rootDir>/src"
+    ],
+    "transform": {
+      "^.+\\.tsx?$": "ts-jest"
+    },
+    "testRegex": "(/__tests__/.*|(\\.|/)(test|spec))\\.tsx?$",
+    "moduleFileExtensions": [
+      "ts",
+      "js",
+      "json"
+    ]
   },
   "dependencies": {
     "axios": "^0.19.0",
@@ -59,17 +61,19 @@
     "pkg-up": "^3.1.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.4.4",
-    "@babel/core": "^7.4.5",
-    "@babel/plugin-proposal-class-properties": "^7.4.4",
-    "@babel/preset-env": "^7.4.5",
+    "@types/find-cache-dir": "^2.0.0",
+    "@types/jest": "^24.0.15",
+    "@types/node": "^12.0.8",
+    "@types/webpack": "^4.4.32",
+    "@typescript-eslint/eslint-plugin": "^1.10.2",
+    "@typescript-eslint/parser": "^1.10.2",
     "axios-mock-adapter": "^1.16.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.8.0",
     "codecov": "^3.5.0",
     "jest": "^24.8.0",
     "standard": "^12.0.1",
-    "webpack": "^4.33.0",
+    "ts-jest": "^24.0.2",
+    "typescript": "^3.5.2",
+    "webpack": "^4.34.0",
     "webpack-cli": "^3.3.4"
   }
 }
