@@ -1,10 +1,8 @@
-'use strict'
-
-import Options from '../../src/Options'
+import Options from './Options'
 
 test('Can initialize with default options', () => {
   const options = new Options()
-  expect(options.get()).toEqual(Options.defaultOptions)
+  expect(options.get()).toEqual(new Options().get())
 })
 
 test('Can overwrite default options object', () => {
@@ -31,7 +29,7 @@ test('Can overwrite default options object', () => {
 })
 
 test('Can load configuration from json file', () => {
-  const options = new Options('__mocks__/options.json')
+  const options = new Options('src/__mocks__/options.json')
   expect(options.get()).toEqual({
     fonts: [],
     formats: [],
@@ -45,7 +43,7 @@ test('Can load configuration from json file', () => {
 })
 
 test('Can load nested configuration from json file', () => {
-  const options = new Options('__mocks__/optionsNested.json')
+  const options = new Options('src/__mocks__/optionsNested.json')
   expect(options.get()).toEqual({
     fonts: [],
     formats: [],
@@ -59,7 +57,7 @@ test('Can load nested configuration from json file', () => {
 })
 
 test('Can load nested configuration from json file while not being the first', () => {
-  const options = new Options('__mocks__/optionsNestedButNotFirst.json')
+  const options = new Options('src/__mocks__/optionsNestedButNotFirst.json')
   expect(options.get()).toEqual({
     fonts: [],
     formats: [],
@@ -73,12 +71,12 @@ test('Can load nested configuration from json file while not being the first', (
 })
 
 test('Will fallback to default options if it\'s not present in json file', () => {
-  const options = new Options('__mocks__/optionsMissing.json')
-  expect(options.get()).toEqual(Options.defaultOptions)
+  const options = new Options('src/__mocks__/optionsMissing.json')
+  expect(options.get()).toEqual(new Options().get())
 })
 
 test('Can load configuration from neon file', () => {
-  const options = new Options('__mocks__/options.neon')
+  const options = new Options('src/__mocks__/options.neon')
   expect(options.get()).toEqual({
     fonts: {},
     formats: {},
@@ -92,7 +90,7 @@ test('Can load configuration from neon file', () => {
 })
 
 test('Can load configuration from nested neon file', () => {
-  const options = new Options('__mocks__/optionsNested.neon')
+  const options = new Options('src/__mocks__/optionsNested.neon')
   expect(options.get()).toEqual({
     fonts: {},
     formats: {},
@@ -106,7 +104,7 @@ test('Can load configuration from nested neon file', () => {
 })
 
 test('Can load nested configuration from json file while not being the first', () => {
-  const options = new Options('__mocks__/optionsNestedButNotFirst.neon')
+  const options = new Options('src/__mocks__/optionsNestedButNotFirst.neon')
   expect(options.get()).toEqual({
     fonts: {},
     formats: {},
@@ -120,6 +118,6 @@ test('Can load nested configuration from json file while not being the first', (
 })
 
 test('Will fallback to default options if it\'s not present in neon file', () => {
-  const options = new Options('__mocks__/optionsMissing.neon')
-  expect(options.get()).toEqual(Options.defaultOptions)
+  const options = new Options('src/__mocks__/optionsMissing.neon')
+  expect(options.get()).toEqual(new Options().get())
 })
